requirements.txt
speedtest-cli==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
pysnmp==4.4.12
pyfzf==0.3.1